@@ -1,13 +1,18 @@
 import { async, TestBed } from '@angular/core/testing'
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations'
 import { AppComponent } from './app.component'
 import { CurrentWeatherComponent } from './current-weather/current-weather.component'
 import { WeatherService } from './weather/weather.service'
 import { WeatherServiceFake } from './weather/weather.service.fake'
+import { MaterialModule } from './material.module'
+import { CitySearchComponent } from './city-search/city-search.component';
 describe('AppComponent', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [AppComponent, CurrentWeatherComponent],
+      declarations: [AppComponent, CurrentWeatherComponent, CitySearchComponent],
       providers: [{ provide: WeatherService, useClass: WeatherServiceFake }],
+      imports: [FormsModule, ReactiveFormsModule, MaterialModule, NoopAnimationsModule],
     }).compileComponents()
   }))
   it('should create the app', async(() => {
@@ -15,10 +20,10 @@ describe('AppComponent', () => {
     const app = fixture.debugElement.componentInstance
     expect(app).toBeTruthy()
   }))
-  it('should render title in a h1 tag', async(() => {
+  it('should render title in a span tag', async(() => {
     const fixture = TestBed.createComponent(AppComponent)
     fixture.detectChanges()
     const compiled = fixture.debugElement.nativeElement
-    expect(compiled.querySelector('h1').textContent).toContain('LocalCast Weather')
+    expect(compiled.querySelector('span').textContent).toContain('LocalCast Weather')
   }))
 })
