@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core'
+import { Component, OnInit, Input } from '@angular/core'
 import { ICurrentWeather } from '../interfaces'
 import { WeatherService } from '../weather/weather.service'
 
@@ -8,8 +8,7 @@ import { WeatherService } from '../weather/weather.service'
   styleUrls: ['./current-weather.component.css'],
 })
 export class CurrentWeatherComponent implements OnInit {
-  current: ICurrentWeather
-
+  @Input() current: ICurrentWeather
   constructor(private weatherService: WeatherService) {
     this.current = {
       city: '',
@@ -22,8 +21,15 @@ export class CurrentWeatherComponent implements OnInit {
   }
 
   ngOnInit() {
-    this.weatherService
-      .getCurrentWeather('Bethesda', 'US')
-      .subscribe(data => (this.current = data))
+    if (this.weatherService.currentWeather) {
+      this.weatherService.currentWeather.subscribe(data => (this.current = data))
+    }
+  }
+
+  getOrdinal(date: number) {
+    const n = new Date(date).getDate()
+    return n > 0
+      ? ['th', 'st', 'nd', 'rd'][(n > 3 && n < 21) || n % 10 > 3 ? 0 : n % 10]
+      : ''
   }
 }
